Docs, JVP throw=True, and complex-dtype determinant coverage
- README: add determinant/slogdet to the feature list and a short example.
- FAQ: drop determinants from the "other linear-algebra operations" entry now
  that lineax supports them (they no longer belong in the redirect-to-jax list).
- slogdet JVP: use throw=True for the internal tangent solves, matching
  `_linear_solve_jvp`'s convention (a failed tangent solve has nowhere to pipe
  an error, so raise rather than silently return a nan gradient). Pseudoinverse
  solvers never raise here, so the pseudodeterminant path is unaffected; add a
  singular-Hermitian gradient test guarding this.
- Add complex128 coverage (`COMPLEX_DET_CASES`, complex JVP test). This caught
  three real bugs, now fixed:
    * QR.slogdet: `taus * v_norms_sq` mixed complex/float under strict dtype
      promotion; cast the real norms to the complex dtype.
    * Cholesky/HEVD.slogdet: returned a real-dtyped sign for a complex Hermitian
      operator; take the operator's dtype so sign is complex (+/-1+0j), matching
      numpy.linalg.slogdet and the other solvers.
    * determinant(): `sign * exp(lad)` mixed complex/float under strict promotion;
      cast exp(lad) to sign's dtype.

+++ lineax/_solver/cholesky.py
@@ -101,7 +101,9 @@ class Cholesky(AbstractDirectLinearSolver[_CholeskyState]):
         factor, is_nsd = state
         n = factor.shape[0]
         sign_val = -1 if (is_nsd.value and n % 2 == 1) else 1
-        sign = jnp.array(sign_val, dtype=factor.real.dtype)
+        # `sign` is always real (+/-1), but takes the operator's dtype so a complex
+        # (Hermitian) operator yields a complex `sign`, matching `numpy.linalg.slogdet`.
+        sign = jnp.array(sign_val, dtype=factor.dtype)
         lad = 2.0 * jnp.sum(jnp.log(jnp.abs(jnp.diag(factor))))
         return sign, lad
